Sat h1

+++ JafuChatPython/ingest.py
@@ -69,15 +69,15 @@ from multiprocessing import Pool
 from tqdm import tqdm
 
 # Import LangChain libraries for document loading, text splitting, embeddings, and vector storage
-from langchain.document_loaders import (
+from langchain_community.document_loaders import (
     CSVLoader, EverNoteLoader, PyMuPDFLoader, TextLoader,
     UnstructuredEmailLoader, UnstructuredEPubLoader, UnstructuredHTMLLoader,
     UnstructuredMarkdownLoader, UnstructuredODTLoader, UnstructuredPowerPointLoader,
     UnstructuredWordDocumentLoader,
 )
 from langchain.text_splitter import RecursiveCharacterTextSplitter
-from langchain.vectorstores import Chroma
-from langchain.embeddings import HuggingFaceEmbeddings
+from langchain_community.vectorstores import Chroma
+from langchain_community.embeddings import HuggingFaceEmbeddings
 from langchain.docstore.document import Document
 from constants import CHROMA_SETTINGS
 
@@ -186,7 +186,8 @@ def ingest_files(source_directory, persist_directory):
 def rebuild_shelf(shelf):
     src = os.path.join(get_root_dir(), shelf)
     db = get_db(shelf)
-    shutil.rmtree(db)
+    if os.path.exists(db):
+        shutil.rmtree(db)
     print("building db....")
     ingest_files(src, db)
     print("done")
